requirement.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-appconf==1.0.6
django-cors-headers==4.4.0
django-cryptography @ git+https://github.com/saurav-codes/django-cryptography@847cade2cc07af8316d59dc0ded709dee708cbc8
django-rest-knox==5.0.1
djangorestframework==3.15.2
pycparser==2.22
sqlparse==0.5.1
typing_extensions==4.12.2
